Cleanup, fix MySQLSchemaJdbcSqlConnectorTest [HZ-2353]
