+++ hazelcast-sql/src/test/java/com/hazelcast/jet/sql/impl/connector/jdbc/mysql/MySQLMappingJdbcSqlConnectorTest.java
@@ -17,10 +17,17 @@
 package com.hazelcast.jet.sql.impl.connector.jdbc.mysql;
 
 import com.hazelcast.jet.sql.impl.connector.jdbc.MappingJdbcSqlConnectorTest;
+import com.hazelcast.sql.HazelcastSqlException;
+import com.hazelcast.sql.impl.QueryException;
 import com.hazelcast.test.annotation.NightlyTest;
 import com.hazelcast.test.jdbc.MySQLDatabaseProvider;
 import org.junit.BeforeClass;
+import org.junit.Test;
 import org.junit.experimental.categories.Category;
+
+import java.sql.SQLException;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 @Category(NightlyTest.class)
 public class MySQLMappingJdbcSqlConnectorTest extends MappingJdbcSqlConnectorTest {
@@ -28,5 +35,21 @@ public class MySQLMappingJdbcSqlConnectorTest extends MappingJdbcSqlConnectorTes
     @BeforeClass
     public static void beforeClass() {
         initialize(new MySQLDatabaseProvider());
+    }
+
+    @Test
+    public void createMappingFails_invalid_externalNameFullName() throws SQLException {
+        createTable(tableName);
+
+        // Create invalid mapping
+        assertThatThrownBy(() -> execute(
+                "CREATE MAPPING " + tableName + " EXTERNAL NAME " + "foo.public.bar" + tableName + " ("
+                        + " id INT, "
+                        + " name VARCHAR "
+                        + ") "
+                        + "DATA CONNECTION " + TEST_DATABASE_REF
+        )).isInstanceOf(HazelcastSqlException.class)
+          .hasRootCauseInstanceOf(QueryException.class)
+          .hasStackTraceContaining("Invalid external name");
     }
 }
